make documents and fix qr_code

+++ app/Console/Commands/deleteFiles.php
@@ -2,6 +2,7 @@
 
 namespace App\Console\Commands;
 
+use App\Models\Deceased;
 use App\Models\Image;
 use App\Models\Video;
 use Illuminate\Console\Command;
@@ -31,12 +32,13 @@ class deleteFiles extends Command
         $files = [
             ...Storage::disk('public')->files('images'),
             ...Storage::disk('public')->files('videos'),
+            ...Storage::disk('public')->files('images/qrCode'),
         ];
 
         foreach ($files as $value) {
             $strPath = 'storage/' . $value;
 
-            if (! Image::where('path', $strPath)->exists() && ! Video::where('path', $strPath)->exists()) {
+            if (! Image::where('path', $strPath)->exists() && ! Video::where('path', $strPath)->exists() && ! Deceased::where('qr_code', $strPath)->exists()) {
                 Storage::delete(str_replace('storage', 'public', $strPath));
             }
